+++ scale.h
@@ -1,5 +1,3 @@
-
-
 void init_scale();
-int read_scale();
+float read_scale();
 void tare_scale();
